multiagente-cpc/pyproject.toml
[tool.poetry]
name = "multiagente_cpc"
version = "0.1.0"
description = ""
authors = ["username <author@mercadolibre.com>"]
classifiers = [
  "Environment :: Web Environment",
  "Intended Audience :: Developers",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development",
]


[tool.poetry.dependencies]
python = ">=3.10,<4"
flask = "^3.1.1"
flask-restx = "^1.3.0"
importlib-metadata = "<8"
melitk-otel = "^0.6.0"
newrelic = "^10.12.0"
opentelemetry-instrumentation-flask = "^0.54b1"
opentelemetry-instrumentation-wsgi = "^0.54b1"
opentelemetry-api = "^1.20.0"
opentelemetry-sdk = "^1.20.0"
pinecone = "^7.0.0"
google-genai = "^1.0.0"
protobuf = ">=5.29.6"
# melitk-secrets — available in Fury runtime, no need to declare here

[tool.poetry.group.dev.dependencies]
bandit = "^1.8.3"
black = "^25.1.0"
flake8 = "^7.2.0"
mypy = "^1.15.0"
pre-commit = "^4.2.0"
pylint = "^3.3.7"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
pytest-flask = "^1.3.0"


[[tool.poetry.source]]
name = "meli"
url = "https://pypi.artifacts.furycloud.io/simple/"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
